requirements.txt
# Core
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Image Processing
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# OCR
google-cloud-vision==3.5.0
pytesseract==0.3.10

# Fuzzy Matching
rapidfuzz==3.6.1
python-Levenshtein==0.23.0

# Frontend
streamlit==1.30.0
requests==2.31.0

# Dev
pytest==7.4.4
black==24.1.1
